Moved the solution and added model files.
